% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adata2sce.R
\name{adata2sce}
\alias{adata2sce}
\title{Coverts AnnData object from Python to SingleCellExperiment object in R.}
\usage{
adata2sce(AnnData)
}
\arguments{
\item{AnnData}{A Python adata.AnnData object.}
}
\value{
A SingleCellExperiment object
}
\description{
The AnnData object here can be read from .h5ad file with the help of 
`reticulate`.
}
\examples{
sce <- adata2sce(adata)
}
