% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_clusters_from_AnnData.R
\name{remove_clusters_from_AnnData}
\alias{remove_clusters_from_AnnData}
\title{Inplace Remove Small Outlier Groups from AnnData}
\usage{
remove_clusters_from_AnnData(AnnData, clusters, ToRemove,
  plotCheck = FALSE, plotCheckMethod = "umap")
}
\arguments{
\item{AnnData}{A Python anndata.AnnData object.}

\item{clusters}{Array of int. The array that `plot_AnnData_UMAP_2D(adata, label_by_cluster = TRUE)` returns}

\item{ToRemove}{numeric. Can be an array of numbers or just one number that 
specify the small groups in the plot. `c(1, 2)`, `15L`, `c('1', 2)` are all
acceptable}

\item{plotCheck}{Boolean, default FALSE. Whether to directly plot the 
filtered AnnData to have a check. The coloring follows the originial labels 
from the AnnData.}

\item{plotCheckMethod}{Character, default "umap", can only be chosen from 
"umap" and "tsne". This would serve as the `method` argument when this 
function internally calls `plot_AnnData_UMAP_2D(adata, method = plotCheckMethod)`.}
}
\description{
Inplace Remove Small Outlier Groups from AnnData
}
\examples{
clusters <- plot_AnnData_UMAP_2D(adata, label_by_cluster = TRUE, 
                                 n_cluster = 30)
remove_clusters_from_AnnData(adata, clusters, c(18, 29), plotCheck = TRUE)
}
