% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce2adata.R
\name{sce2adata}
\alias{sce2adata}
\title{Coverts SingleCellExperiment object from R to anndata.AnnData object in Python}
\usage{
sce2adata(SCE, assay = "counts", save_h5ad = FALSE,
  h5ad_filename = NULL)
}
\arguments{
\item{SCE}{A SingleCellExperiment object.}

\item{assay}{Character, optional, default "counts". The name of assay of
interest that will be set as the primary matrix of the output AnnData.
Available options can be listed by `names(SCE@assays$data)`}

\item{save_h5ad}{Boolean, optional, default FALSE. Whether to save the
converted AnnData object to file.}

\item{h5ad_filename}{character, optional, default NULL.}
}
\value{
A Python anndata.AnnData object
}
\description{
The AnnData object here can be saved to .h5ad file and read into Python
interactive console.
}
\examples{
adata <- adata2sce(sce)
}
