% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AnnData_UMAP_2D.R
\name{plot_AnnData_UMAP_2D}
\alias{plot_AnnData_UMAP_2D}
\title{Plot a 2D UMAP in the same way that STREAM plots.}
\usage{
plot_AnnData_UMAP_2D(AnnData)
}
\arguments{
\item{AnnData}{A Python anndata.AnnData object that has 
anndata.AnnData.obsm['X_vis_umap'] value}
}
\description{
Plot a 2D UMAP in the same way that STREAM plots.
}
\examples{
plot_AnnData_UMAP_2D(adata)
}
